Initial push with changes to meson build
This consists of the first changes to move to the meson build system,
and also to compile both libifisam and libvbisam libraries for use as
alternative ISAM implementations.

The verison of libvbisam contained within the repository is based on
the version found contained within GnuCOBOL, and is 2.1.1 at the time.