Initial push of changes for extra definitions
This change includes a new way of determining the field information from a .fld
file that is used by the legacy SuperNova system, now owned by MicroFocus. Like
CISAM itself, no actual field names are supplied in the file, so default ones
are generated of the 'fldNNN' where NNN is an incrementing number starting from
1. Likewise, no index information is provided, it is picked up when the first
record is read from the package -- this is only provided by the ISAMtable
object which adds an extra layer of logic above the ISAMisam object which only
provides lowlevel ISAM calls.

Also reorganised the backend logic to make more use of importlib.import_module
which directly pulls in the appropriate support module, removing the need for
logic in the __init__.py files, though the code is still there commented out
with NOT USED..END NOT USED comments.

A further test, test_40, has been created which will use the new fldfile
module within the tabdefns sub-package. test_11, has been modified to make
use of an iterator class to perform the record retrieval, which also provides
the conversion of the error raised when the end of file is reached into the
StopIteration exception expected by the for statement.