Fix problem with transaction logging
This fixes the problem with using the VBISAM backend, where the
index file keeps getting exteneded with logging information, even
when there are no updates being done on the file.

This was tracked down to the underlying VBISAM library not being
initialised correctly (or at all) leading the transaction logging
file handle being set to zero, which then became the same as the
first file opened during an isbuild/isopen call.

The vbisam handling modules now call the vb_get_rtd() routine to
explicitly initialise the vbisam library.