Revert back to vb_get_rtd
This reverses the change to vb_init_rtd back to vb_get_rtd to initialise
the underlying VBISAM library as this function calls vb_init_rtd the first
time through. Although the pyisam package does not store the return value,
this could in future be used to access the global variables such as iserrno
directly.