Fix for issue #2, test_12 failing
This commit makes a change to the read() method of the ISAMtable class which
means that the previously failing test will now work as originally expected.

The underlying ISAM table will be automatically opened if not already using the
mode and lock options that were specified when the ISAMtable was created.

There is a comment block within the read() method that explains how the arguments
are processed.